+++ src/storm-conv/settings/modules/JaniExportSettings.cpp
@@ -24,6 +24,7 @@ namespace storm {
             const std::string JaniExportSettings::eliminateFunctionsOptionName = "remove-functions";
             const std::string JaniExportSettings::replaceUnassignedVariablesWithConstantsOptionName = "replace-unassigned-vars";
             const std::string JaniExportSettings::simplifyCompositionOptionName = "simplify-composition";
+            const std::string JaniExportSettings::performLocationElimination = "location-elimination";
             
             JaniExportSettings::JaniExportSettings() : ModuleSettings(moduleName) {
                 this->addOption(storm::settings::OptionBuilder(moduleName, locationVariablesOptionName, true, "Variables to export in the location").addArgument(storm::settings::ArgumentBuilder::createStringArgument("variables", "A comma separated list of automaton and local variable names seperated by a dot, e.g. A.x,B.y.").setDefaultValueString("").build()).build());
@@ -36,6 +37,10 @@ namespace storm {
                 this->addOption(storm::settings::OptionBuilder(moduleName, eliminateFunctionsOptionName, false, "If set, transforms the model such that functions are eliminated.").build());
                 this->addOption(storm::settings::OptionBuilder(moduleName, replaceUnassignedVariablesWithConstantsOptionName, false, "If set, local and global variables that are (a) not assigned to some value and (b) have a known initial value are replaced by constants.").build());
                 this->addOption(storm::settings::OptionBuilder(moduleName, simplifyCompositionOptionName, false, "If set, attempts to simplify the system composition.").build());
+                this->addOption(storm::settings::OptionBuilder(moduleName, performLocationElimination, false, "If set, location elimination will be performed before the model is built.").setIsAdvanced()
+                                    .addArgument(storm::settings::ArgumentBuilder::createUnsignedIntegerArgument("location-heuristic", "If this number of locations is reached, no further unfolding will be performed").setDefaultValueUnsignedInteger(10).makeOptional().build())
+                                    .addArgument(storm::settings::ArgumentBuilder::createUnsignedIntegerArgument("edges-heuristic", "Determines how many new edges may be created by a single elimination").setDefaultValueUnsignedInteger(10000).makeOptional().build())
+                                    .build());
             }
             
             bool JaniExportSettings::isAllowEdgeAssignmentsSet() const {
@@ -92,6 +97,18 @@ namespace storm {
             
             bool JaniExportSettings::isSimplifyCompositionSet() const {
                 return this->getOption(simplifyCompositionOptionName).getHasOptionBeenSet();
+            }
+            
+            bool JaniExportSettings::isLocationEliminationSet() const {
+                return this->getOption(performLocationElimination).getHasOptionBeenSet();
+            }
+            
+            uint64_t JaniExportSettings::getLocationEliminationLocationHeuristic() const{
+                return this->getOption(performLocationElimination).getArgumentByName("location-heuristic").getValueAsUnsignedInteger();
+            }
+            
+            uint64_t JaniExportSettings::getLocationEliminationEdgesHeuristic() const {
+                return this->getOption(performLocationElimination).getArgumentByName("edges-heuristic").getValueAsUnsignedInteger();
             }
             
             void JaniExportSettings::finalize() {